UserInterface/requirements.txt
asgiref==3.8.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
docker==7.1.0
Flask==3.0.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
numpy==2.1.2
requests==2.32.3
scipy==1.14.1
sqlparse==0.5.1
urllib3==2.2.3
Werkzeug==3.0.4
